Extract exception handling query to its own file
